+++ src/js/variables.js
@@ -1,0 +1,15 @@
+const form = document.querySelector('.js__form');
+const namePreview = document.querySelector('.js__preview_name');
+const jobPreview = document.querySelector('.js__preview_job');
+const phonePreview = document.querySelector('.js__preview_phone');
+const emailPreview = document.querySelector('.js__preview_email');
+const linkedinPreview = document.querySelector('.js__preview_linkedn');
+const githubPreview = document.querySelector('.js__preview_github');
+const cardPreview = document.querySelector('.js__preview');
+const regexPhone = document.querySelector('.js-msg-phone');
+const regexEmail = document.querySelector('.js-msg-email');
+const regexLinkedin = document.querySelector('.js-msg-linkedin');
+const regexGit = document.querySelector('.js-msg-git');
+const expresionEmail = /^[\w-\.]+@([\w-]+\.)+[\w-]{2,4}$/;
+const expresionUrl = /[(http(s)?):\/\/(www\.)?a-zA-Z0-9@:%._\+~#=]{2,256}\.[a-z]{2,6}\b([-a-zA-Z0-9@:%_\+.~#?&//=]*)/;
+const expressionPhone = /(?:([+]\d{1,4})[-.\s]?)?(?:[(](\d{1,3})[)][-.\s]?)?(\d{1,4})[-.\s]?(\d{1,4})[-.\s]?(\d{1,9})/g;
